refactor: extract CLI systemd lifecycle to TypeScript
